Krishi Mitra/Back-end/requirements.txt
Flask==2.0.1

Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Flask-WTF==0.15.1
bcrypt==3.2.0
python-dotenv==0.19.0
requests==2.26.0
Pillow==8.3.2
tensorflow==2.6.0
numpy==1.19.5
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
models==0.1.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-CORS==3.0.10
requests==2.26.0
Werkzeug==2.0.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
bcrypt==3.2.0
python-dotenv==0.19.0
Pillow==8.3.2
tensorflow==2.6.0
numpy==1.19.5
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.3
gunicorn==20.1.0
psycopg2-binary==2.9.1  
models==0.1.0   
Flask==2.0.1
    Flask-SQLAlchemy==2.5.1
    Flask-Login==0.5.0
    Flask-CORS==3.0.10
    requests==2.26.0
    Werkzeug==2.0.1
    Flask-JWT-Extended==4.3.1
    Flask-Migrate==3.1.0  
    bcrypt==3.2.0
    python-dotenv==0.19.0
    Pillow==8.3.2
    tensorflow==2.6.0
    numpy==1.19.5
    pandas==1.3.3
    Flask-SQLAlchemy==2.5.1